requirements.txt
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.7
django-heroku==0.3.1
django-tinymce==3.3.0
gunicorn==20.1.0
psycopg2==2.9.1
pycodestyle==2.7.0
python-dotenv==0.19.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.3.0
